Added .md files to respective folders
