+++ components/BlogPage.js
@@ -9,7 +9,6 @@ function BlogPage(props) {
 
     return (
         <SubContainer>
-        {/* <div className="border-2 border-pacific-800 rounded-lg block my-8 p-1 lg:p-4 xl:p-4"> */}
             <div className="m-6">
                 <div className="text-center">
                     <Link href={sectionLink}>
@@ -28,7 +27,6 @@ function BlogPage(props) {
                 </article>
             </div>
             </SubContainer>
-        // </div>
     )
 }
