+++ package.json
@@ -9,8 +9,9 @@
   },
   "dependencies": {
     "autoprefixer": "^10.1.0",
+    "glob": "^7.1.6",
     "gray-matter": "^4.0.2",
-    "next": "^10.0.4-canary.4",
+    "next": "^10.0.4-canary.5",
     "next-images": "^1.6.2",
     "npm": "^6.14.9",
     "postcss": "^8.2.1",
